app/auth/dependencies.py
from typing import Union
from datetime import datetime, timedelta
from jose import jwt

from app.config.setting import Setting
from app.users.dependencies import get_user_from_db_active

pwd_context = Setting.pwd_context

def get_password_hash(password):
    return pwd_context.hash(password)

def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)

async def authenticate_user(username: str, password: str):
    user = await get_user_from_db_active(username)
    if not user:
        return False
    if not verify_password(password, user['hashed_password']):
        return False
    return user

def create_access_token(data: dict, expires_delta: Union[timedelta, None] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now() + expires_delta
    else:
        expire = datetime.now() + timedelta(minutes=Setting.ACCESS_TOKEN_EXPIRE_MINUTES)
        
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, Setting.SECRET_KEY, algorithm=Setting.ALGORITHM)
    return encoded_jwt
